Fix: pylint issues with MyApp, Remove python 3.8,9, fix pylint.yml
Because wx.App is a C++ class, it uses CamelCase for functions.
Because OnInit is a "init-alike" callback, defining attributes in there
  is expected.

Using a slotted dataclass requires 3.10+.  Better do that than
"allow" or have to check against adding gamesettings that don't exist.

Fail on bad pylint score, or any Warning, Error, or Fatal.