Fix: Go to invisible breaks repeatable
For invisible breaks, "Go to break" should just mean "add N minutes
to the clock."  Because it wasn't added to the list of unscheduled
breaks, this could be done repeatedly in the same round.

Therefore if one TD is tasked to "add the break", and another has
on their own initiative already done so, double break would be added
without warning.  Yes, that's what currently happens, but why
propagate deliberately?