Add:accelerator dataclass, shortcut keys.
Created dataclass to contain information about a
shortcut key/context menu entry.  Added the information in __init__
subfunction as a list, and changed _create_context_menu to
_create_shortcuts.

Thought: maybe put flag in for "create accelerator/create menu item"
booleans in case we want a shortcut key not in the menu or vice versa.

Still has issues, but they (mostly) work.