Add: make sounds work.
The idea will be to have a sound subdirectory that can be filled with
the desired sounds (beeps, spoken lines, etc.) and offer samples.

There will be a default sound that plays if the desired one does not
exist in the pattern used.

Currently, there are multiple triggers that shouldn't happen, such as
when starting a game.