Change: end-of-game behaviour.  Introduce countback. (#50)
ACBL directors will start clocks with one round extra, so that they
can point to the clock and show how late the last tables are.

Just putting up a "GAME OVER" screen or 00:00 doesn't do that, but
"last game" settings are only saved when the game is over and the clock
is finished, so "one round more" would never end cleanly.

Changed the clock to not stop at end of game, so it will now start
counting -00:01, -00:02,... at the "Game Over" screen.  Now directors
can point and say "You're 5 minutes overtime" with the "correct"
clock timing.