Fix: limits passing in BaseV, fail for SoundV
Two issues:
1. Limits weren't being cloned properly, so changed the
format of BaseValidator to store the limits as an object, and
pass it everywhere.  Init for derived validators need to
take limits as an input (from clone) or the default set (for init).

2. Fixed sound testing failure reporting.  New utilities use just
the barename of the mp3 file, and do all the searching for you.  Added
silence.mp3 so that validation of sounds (which happen every time you
click the sound button in preferences, and when you close the box) don't
make noise and annoy the players.